Bug fix: negated existance matchers
I ran into an interesting problem today where the coincidental overlap
of a site prism method with that on a regular class, meant that the
negated existance matcher in some existing code stopped working.

This appears to be because of
https://github.com/natritmeyer/site_prism/pull/390, expecting a
non-standard method to always be present in order to have site prism
wait correctly.

I've added a `respond_to?` check for the negated matcher, and now fall
back to the standard `!positive_matcher` when no such method exists.